# backend/docs/swagger.yaml
basePath: /api/
definitions:
...
  models.Bookmark:
    properties:
      icon:
        description: as in Icon; "" uses the linked host's or service's
        type: string
      id:
        type: integer
      mac:
        description: linked host, "" for none
        type: string
      name:
        type: string
      note:
        description: shown under the name instead of the address
        type: string
      port:
        description: linked service on that host, 0 for the host itself
        type: integer
      url:
        type: string
    type: object
...
  models.Container:
    properties:
      category:
        type: string
      cid:
        description: container ID
        type: string
      connectorID:
        type: integer
      host:
        type: string
      icon:
        description: icon suggestion from the image name
        type: string
      id:
        type: integer
      image:
        type: string
      ip:
        description: IP of the Docker host
        type: string
      mac:
        description: host it runs on, "" if that host is not known
        type: string
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/models.ContainerPort'
        type: array
      project:
        description: compose project
        type: string
      state:
        description: '"running", "exited"...'
        type: string
      status:
        type: string
    type: object
  models.ContainerPort:
    properties:
      ip:
        type: string
      private:
        type: integer
      proto:
        type: string
      public:
        description: 0 when not published on the host
        type: integer
    type: object
...
  models.Guest:
    properties:
      connectorID:
        type: integer
      cpus:
        type: integer
      hostMac:
        description: the guest's own host, "" if not discovered
        type: string
      id:
        type: integer
      ip:
        description: if the hypervisor knows it
        type: string
      mac:
        description: first NIC, lower case
        type: string
      maxMem:
        description: bytes
        type: integer
      name:
        type: string
      node:
        description: hypervisor node name
        type: string
      nodeMac:
        description: the node's host, "" if not known
        type: string
      status:
        description: '"running", "stopped"...'
        type: string
      tags:
        description: as set in the hypervisor, ";" separated
        type: string
      type:
        description: '"qemu" (VM) or "lxc"'
        type: string
      uptime:
        description: seconds
        type: integer
      vmid:
        type: integer
    type: object
...
  models.Icon:
    properties:
      icon:
        type: string
      mac:
        type: string
      port:
        type: integer
    type: object
  models.Item:
    properties:
      bookmark:
        description: bookmark ID
        type: integer
      icon:
        description: unused since bookmarks; hosts and services use the icons table
        type: string
      id:
        type: integer
      kind:
        description: '"host", "service" or "bookmark" ("link" before bookmarks existed)'
        type: string
      mac:
        description: host and service
        type: string
      name:
        description: optional for host and service
        type: string
      port:
        description: service
        type: integer
      sort:
        type: integer
      tag:
        type: string
      url:
        description: overrides where a host or service tile goes
        type: string
    type: object
...
  models.MapHost:
    properties:
      bookmarks:
        description: bookmarks linked to this host or its services
        items:
          $ref: '#/definitions/models.Bookmark'
        type: array
      category:
        description: chosen category, or the suggested one, see package category
        type: string
      containers:
        items:
          $ref: '#/definitions/models.Container'
        type: array
      date:
        type: string
      dns:
        type: string
      guestOf:
        allOf:
        - $ref: '#/definitions/models.Guest'
        description: set when this host is a VM or LXC on a known hypervisor
      guests:
        description: VMs and LXCs when this host is a hypervisor node
        items:
          $ref: '#/definitions/models.Guest'
        type: array
      hasPos:
        description: 'false: the map places the host automatically'
        type: boolean
      hw:
        type: string
      icon:
        description: icon set for the host, "" if none
        type: string
      id:
        type: integer
      iface:
        type: string
      ip:
        type: string
      known:
        type: integer
      mac:
        type: string
      name:
        type: string
      now:
        type: integer
      subnetID:
        description: 0 when the host is in no configured subnet
        type: integer
      suggested:
        description: true when Category is a suggestion, not chosen
        type: boolean
      web:
        description: open ports that serve a web page
        items:
          $ref: '#/definitions/models.Port'
        type: array
      x:
        format: float64
        type: number
      "y":
        format: float64
        type: number
    type: object
...
  models.Port:
    properties:
      category:
        description: category the connector gave the service
        type: string
      container:
        description: container publishing the port
        type: string
      first:
        type: string
      id:
        type: integer
      ip:
        type: string
      last:
        type: string
      mac:
        type: string
      port:
        type: integer
      service:
        type: string
      source:
        description: Set when a connector reported this port
        type: integer
      title:
        description: title of the web page on this port, if any
        type: string
      web:
        description: '"http", "https" or ""'
        type: string
    type: object
...
  models.View:
    properties:
      id:
        type: integer
      layout:
        description: '"tiles" or "map"'
        type: string
      name:
        type: string
      sort:
        type: integer
      tags:
        description: groups in order, comma separated; empty shows every tag
        type: string
    type: object
  models.ViewData:
    properties:
      groups:
        items:
          $ref: '#/definitions/models.ViewGroup'
        type: array
      view:
        $ref: '#/definitions/models.View'
    type: object
  models.ViewGroup:
    properties:
      items:
        items:
          $ref: '#/definitions/models.ViewItem'
        type: array
      tag:
        type: string
    type: object
  models.ViewItem:
    properties:
      bookmark:
        description: bookmark ID
        type: integer
      host:
        $ref: '#/definitions/models.MapHost'
      icon:
        description: unused since bookmarks; hosts and services use the icons table
        type: string
      id:
        type: integer
      kind:
        description: '"host", "service" or "bookmark" ("link" before bookmarks existed)'
        type: string
      link:
        description: where clicking the tile goes ("" opens host details)
        type: string
      mac:
        description: host and service
        type: string
      name:
        description: optional for host and service
        type: string
      port:
        description: service
        type: integer
      sort:
        type: integer
      status:
        description: '"up", "down" or "unknown"'
        type: string
      subtitle:
        type: string
      tag:
        type: string
      title:
        type: string
      url:
        description: overrides where a host or service tile goes
        type: string
    type: object
...
info:
  contact:
    url: https://github.com/vipthomps/Gimle
  description: 'Home lab dashboard: network discovery, a map of hosts, bookmarks,
    views and docs'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gimlé API
  version: 1.0.0
paths:
...
  /icons:
    post:
      consumes:
      - application/json
      description: Port 0 is the host itself. Icon is a dashboard-icons name (e.g.
        "immich"), an image URL, "none", or "" to go back to the default
      parameters:
      - description: Icon
        in: body
        name: icon
        required: true
        schema:
          $ref: '#/definitions/models.Icon'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Icon'
      summary: Set the icon of a host or service
      tags:
      - views
  /items:
    post:
      consumes:
      - application/json
      description: Kind "host" needs Mac, "service" needs Mac and Port. Bookmarks
        are tagged through /bookmarks
      parameters:
      - description: Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.Item'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Item'
      summary: Create or update an item
      tags:
      - views
  /items/{id}:
    delete:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete an item
      tags:
      - views
  /items/{id}/move:
    post:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: -1 or 1
        in: query
        name: dir
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Move an item earlier or later within its tag
      tags:
      - views
...
  /tags/host/{mac}:
    get:
      description: Tag items for a host and its services
      parameters:
      - description: Host MAC
        in: path
        name: mac
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Item'
            type: array
      summary: Tags on a host
      tags:
      - views
...
  /views:
    get:
      description: Custom dashboard tabs in order
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.View'
            type: array
      summary: List views
      tags:
      - views
    post:
      consumes:
      - application/json
      description: ID 0 creates a view at the end. Layout is "tiles" or "map"
      parameters:
      - description: View
        in: body
        name: view
        required: true
        schema:
          $ref: '#/definitions/models.View'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.View'
      summary: Create or update a view
      tags:
      - views
  /views/{id}:
    delete:
      description: Deletes the view only. Tags stay on hosts, services and links
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete a view
      tags:
      - views
    get:
      description: Items come with a title, a link and an up/down status
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ViewData'
      summary: View with groups and items
      tags:
      - views
  /views/{id}/move:
    post:
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: integer
      - description: -1 or 1
        in: query
        name: dir
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Move a view left or right
      tags:
      - views
...
swagger: "2.0"